backend/requirements.txt
flask
flask-cors
pillow
tensorflow
